+++ macro/run_simScript.py
@@ -77,6 +77,13 @@ pg_parser.add_argument(
 # === Genie subcommand ===
 genie_parser = subparsers.add_parser("Genie", help="Genie for reading and processing neutrino interactions")
 genie_parser.add_argument(
+    "--GenieOption",
+    dest="GenieOption",
+    default=1,
+    type=int,
+    help="Genie generation option: (1 standard, 4 GENIE geometry driver)"
+)
+genie_parser.add_argument(
     "--z_start_nu",
     dest="z_start_nu",
     default=2844.2850,
@@ -559,8 +566,10 @@ if options.command == "Genie":
     ut.checkFileExists(inputFile)
     primGen.SetTarget(0.0, 0.0)  # do not interfere with GenieGenerator
     Geniegen = ROOT.GenieGenerator()
+    Geniegen.SetGenerationOption(options.GenieOption - 1) # 0 standard, 3 GENIE geometry driver
     Geniegen.Init(inputFile, options.firstEvent)
-    Geniegen.SetPositions(ship_geo.target.z0, options.z_start_nu, options.z_end_nu)
+    if options.GenieOption == 1:
+     Geniegen.SetPositions(ship_geo.target.z0, options.z_start_nu, options.z_end_nu)
     primGen.AddGenerator(Geniegen)
     options.nEvents = Geniegen.GetNevents() if options.nEvents == -1 else min(options.nEvents, Geniegen.GetNevents())
     run.SetPythiaDecayer("DecayConfigNuAge.C")