+++ CHANGELOG.md
@@ -13,6 +13,8 @@ it in future.
 ## Unreleased
 ### Added
 
+* Added GenieOption to read simulations with the GENIE Geometry Driver (i.e. gevgen_fnal) 
+
 ### Changed
 
 ### Fixed